+++ lib/panoptes/client/me.rb
@@ -1,8 +1,11 @@
+require 'openssl'
+require 'jwt'
+
 module Panoptes
   class Client
     module Me
       def me
-        get("/me")["users"][0]
+        panoptes.get("/me")["users"][0]
       end
     end
   end