+++ lib/panoptes/session.rb
@@ -1,0 +1,6 @@
+module Panoptes
+  class Session
+    def initialize(auth)
+    end
+  end
+end