+++ spec/spec_helper.rb
@@ -4,7 +4,9 @@ require 'panoptes-client'
 
 require 'webmock/rspec'
 require 'vcr'
+require 'timecop'
 require 'pry'
+require 'openssl'
 
 VCR.configure do |config|
   config.cassette_library_dir = 'spec/fixtures/vcr'
@@ -16,34 +18,50 @@ VCR.configure do |config|
   config.filter_sensitive_data("<CLIENT_SECRET") { test_client_secret }
 end
 
-def test_url;           ENV.fetch("ZOONIVERSE_URL",           'https://panoptes-staging.zooniverse.org'); end
-def test_talk_url;      ENV.fetch("ZOONIVERSE_TALK_URL",      'https://talk-staging.zooniverse.org'); end
-def test_access_token;  ENV.fetch("ZOONIVERSE_ACCESS_TOKEN",  'x'*64); end
-def test_client_id;     ENV.fetch("ZOONIVERSE_CLIENT_ID",     'x'*64); end
-def test_client_secret; ENV.fetch("ZOONIVERSE_CLIENT_SECRET", 'x'*64); end
+def fake_keypair
+  @fake_keypair ||= OpenSSL::PKey::RSA.generate 2048
+end
+
+def fake_public_key_path
+  @fake_public_key_path ||= Tempfile.new('panoptes-client-pubkey').tap do |file|
+    file.write(fake_keypair.public_key.to_s)
+    file.close
+  end
+  @fake_public_key_path.path
+end
+
+def fake_access_token
+  JWT.encode({"data" => {"id" => 1323869}}, fake_keypair, 'RS512')
+end
+
+def test_url;           ENV.fetch("ZOONIVERSE_URL",             'https://panoptes-staging.zooniverse.org'); end
+def test_talk_url;      ENV.fetch("ZOONIVERSE_TALK_URL",        'https://talk-staging.zooniverse.org'); end
+def test_access_token;  ENV.fetch("ZOONIVERSE_ACCESS_TOKEN",    fake_access_token); end
+def test_public_key;    ENV.fetch("ZOONIVERSE_PUBLIC_KEY_PATH", fake_public_key_path); end
+def test_client_id;     ENV.fetch("ZOONIVERSE_CLIENT_ID",       'x'*64); end
+def test_client_secret; ENV.fetch("ZOONIVERSE_CLIENT_SECRET",   'x'*64); end
 
 def unauthenticated_client
-  Panoptes::Client.new(url: test_url)
+  Panoptes::Client.new(env: :test)
 end
 
 def application_client
   Panoptes::Client.new(
-    url: test_url,
-    auth_url: test_url,
-    auth: { client_id: test_client_id, client_secret: test_client_secret }
+    env: :test,
+    auth: {client_id: test_client_id, client_secret: test_client_secret}
   )
 end
 
 def user_client
   Panoptes::Client.new(
-    url: test_url,
-    auth_url: test_url,
-    auth: { token: test_access_token }
+    env: :test,
+    auth: {token: test_access_token},
+    public_key_path: test_public_key,
   )
 end
 
 def talk_client
-  Panoptes::TalkClient.new(url: test_talk_url)
+  user_client
 end
 
 def api_url(path)